requirements.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.3.0
Django==1.11.13
django-auditlog==0.4.5
django-cors-headers==2.2.0
django-extensions==2.0.7
django-filter==1.1.0
django-jsonfield==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
idna==2.6
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.14.0
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
